refactor(Blazor): move common components to Common namespace
Move remaining general-use components to the ~Web.Components.Common
namespace. Separates the components reusable in other projects to a single area,
to be extracted and converted into a standalone project later.
