+++ NjordinSight.Web/Components/Common/LocalizableComponent.razor
