+++ NjordinSight.Web/Components/Common/Modal.razor
@@ -14,8 +14,6 @@
             @if(ReadOnly)
             {
                 @ChildContent
-                @*<div>
-                </div>*@
                 <ActionButtonMenu>
                     @foreach(var dialogOption in _dialogResultButtons)
                     {
