+++ NjordinSight.Web/Components/Common/SearchInputTable.razor
@@ -6,7 +6,6 @@
     <CascadingValue Value="@ErrorMessage">
         <ErrorMessageBox />
     </CascadingValue>
-    @*<label hidden=@SearchIsValid class="alert-info">@InvalidSearchMessage</label>*@
     <ThemedComponent>
         <EditForm Model="@this">
             <table class="table search-parameter-table">
