+++ NjordinSight.Web/Components/Common/Strings.resx
