+++ NjordinSight.Web/Components/ThemedComponent.razor.css
@@ -204,12 +204,6 @@
             color: var(--theme-text-color-0, white) !important;
         }
 
-/*::deep :focus-visible {
-    // TODO: Find out why this is not applied unless written 
-    // using the browswer dev tools.
-    outline: 0.1rem solid var(--theme-hexcolor-secondary-2-2, blue);
-}*/
-
 ::deep .form-control:focus {
     box-shadow: 0 0 0 0.2rem var(--theme-hexcolor-secondary-2-2, blue);
 }
